change structure and CR applied
